Change prefix of AwsDynamoDB hook module (#11209)
* align import path of AwsDynamoDBHook in aws providers
