+++ tests/providers/amazon/aws/hooks/test_dynamodb.py
@@ -20,7 +20,7 @@
 import unittest
 import uuid
 
-from airflow.providers.amazon.aws.hooks.aws_dynamodb import AwsDynamoDBHook
+from airflow.providers.amazon.aws.hooks.dynamodb import AwsDynamoDBHook
 
 try:
     from moto import mock_dynamodb2
